backend/requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
mysqlclient==2.2.5
sqlparse==0.5.1
typing_extensions==4.12.2
python-dotenv
django-filter